@@ -4,11 +4,11 @@
   "private": true,
   "main": "main.js",
   "dependencies": {
-    "@exponent/ex-navigation": "file:../",
-    "@exponent/vector-icons": "~2.0.0",
-    "exponent": "~12.0.3",
+    "@expo/ex-navigation": "file:../",
+    "@expo/vector-icons": "~4.0.0",
+    "expo": "^15.0.2",
     "lodash": "^4.13.1",
-    "react": "~15.3.2",
-    "react-native": "git+https://github.com/exponent/react-native#sdk-12.0.0"
+    "react": "~15.4.2",
+    "react-native": "https://github.com/exponent/react-native/archive/sdk-15.0.0.tar.gz"
   }
 }
