@@ -2,13 +2,9 @@
  * @flow
  */
 
-import React, {
-  Children,
-} from 'react';
-import {
-  StyleSheet,
-  View,
-} from 'react-native';
+import React, { Children } from 'react';
+import { StyleSheet, View } from 'react-native';
+import PropTypes from 'prop-types';
 import DrawerLayout from 'react-native-drawer-layout-polyfill';
 import PureComponent from '../utils/PureComponent';
 import StaticContainer from 'react-static-container';
@@ -56,7 +52,6 @@ export class ExNavigationDrawerContext extends ExNavigatorContext {
   }
 }
 
-
 type Props = {
   id: string,
   navigatorUID: string,
@@ -69,7 +64,10 @@ type Props = {
   children: Array<React.Element<any>>,
   drawerPosition?: 'left' | 'right',
   navigation: ExNavigationContext,
-  onRegisterNavigatorContext: (navigatorUID: string, navigatorContext: ExNavigationDrawerContext) => void,
+  onRegisterNavigatorContext: (
+    navigatorUID: string,
+    navigatorContext: ExNavigationDrawerContext
+  ) => void,
   onUnregisterNavigatorContext: (navigatorUID: string) => void,
   navigationState: Object,
 };
@@ -100,12 +98,12 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
   };
 
   static contextTypes = {
-    parentNavigatorUID: React.PropTypes.string,
+    parentNavigatorUID: PropTypes.string,
   };
 
   static childContextTypes = {
-    parentNavigatorUID: React.PropTypes.string,
-    navigator: React.PropTypes.instanceOf(ExNavigationDrawerContext),
+    parentNavigatorUID: PropTypes.string,
+    navigator: PropTypes.instanceOf(ExNavigationDrawerContext),
   };
 
   getChildContext() {
@@ -152,15 +150,14 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
       drawerPosition: this.props.drawerPosition,
       width: this.props.drawerWidth,
       renderNavigationView: this.props.renderNavigationView,
-      style: [
-        this.props.drawerStyle,
-      ],
+      style: [this.props.drawerStyle],
     };
-
 
     return (
       <ExNavigationDrawerLayout
-        ref={component => { this._drawerLayout = component; }}
+        ref={component => {
+          this._drawerLayout = component;
+        }}
         {...drawerLayoutProps}
       />
     );
@@ -191,7 +188,7 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
       <View
         key={drawerItem.id}
         removeClippedSubviews={!isSelected}
-        style={[styles.itemContentInner, {opacity: isSelected ? 1 : 0}]}
+        style={[styles.itemContentInner, { opacity: isSelected ? 1 : 0 }]}
         pointerEvents={isSelected ? 'auto' : 'none'}>
         <StaticContainer shouldUpdate={isSelected}>
           {drawerItem.element}
@@ -205,19 +202,25 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
 
     this._registerNavigatorContext();
 
-    this.props.navigation.dispatch(Actions.setCurrentNavigator(
-      this.state.navigatorUID,
-      this.state.parentNavigatorUID,
-      'drawer',
-      {},
-      [{
-        key: this.props.initialItem,
-      }],
-    ));
+    this.props.navigation.dispatch(
+      Actions.setCurrentNavigator(
+        this.state.navigatorUID,
+        this.state.parentNavigatorUID,
+        'drawer',
+        {},
+        [
+          {
+            key: this.props.initialItem,
+          },
+        ]
+      )
+    );
   }
 
   componentWillUnmount() {
-    this.props.navigation.dispatch(Actions.removeNavigator(this.state.navigatorUID));
+    this.props.navigation.dispatch(
+      Actions.removeNavigator(this.state.navigatorUID)
+    );
     this.props.onUnregisterNavigatorContext(this.state.navigatorUID);
   }
 
@@ -228,7 +231,10 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
 
     if (nextProps.navigationState !== this.props.navigationState) {
       this.setState({
-        renderedItemKeys: this._updateRenderedItemKeys(nextProps, this.state.renderedItemKeys),
+        renderedItemKeys: this._updateRenderedItemKeys(
+          nextProps,
+          this.state.renderedItemKeys
+        ),
       });
     }
   }
@@ -243,7 +249,9 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
     if (prevProps.navigationState !== this.props.navigationState) {
       const navigationState = this.props.navigationState;
       const currentItemKey = navigationState.routes[navigationState.index].key;
-      const navigatorUIDForItemKey = this._getNavigatorContext().getNavigatorUIDForItemKey(currentItemKey);
+      const navigatorUIDForItemKey = this._getNavigatorContext().getNavigatorUIDForItemKey(
+        currentItemKey
+      );
       if (navigatorUIDForItemKey) {
         this.props.navigation.dispatch(
           Actions.setCurrentNavigator(navigatorUIDForItemKey)
@@ -258,7 +266,12 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
     const selectedChild = navState.routes[navState.index];
 
     return [
-      ..._.uniq(_.without([...currentRenderedItemKeys, ...currentDrawerItems], selectedChild.key)),
+      ..._.uniq(
+        _.without(
+          [...currentRenderedItemKeys, ...currentDrawerItems],
+          selectedChild.key
+        )
+      ),
       selectedChild.key,
     ];
   }
@@ -271,7 +284,7 @@ class ExNavigationDrawer extends PureComponent<any, Props, State> {
 
       invariant(
         child.type === ExNavigationDrawerItem,
-        'All children of DrawerNavigation must be DrawerNavigationItems.',
+        'All children of DrawerNavigation must be DrawerNavigationItems.'
       );
 
       const drawerItemProps = child.props;
